paymentScreen and ReceiptScreen.xml Theme
