+++ aumet_custom/aumet_pos_multi_uom_price/i18n/es.po
