+++ aumet_custom/aumet_pos_multi_uom_price/static/description/POS_multi_uom_price.png
