+++ aumet_custom/aumet_pos_multi_uom_price/static/description/icon.png
