--- aumet_custom/pos_multi_uom_price/static/src/xml/multi_uom_price.xml
@@ -1,12 +1,0 @@
-<?xml version="1.0" encoding="UTF-8"?>
-<templates id="template" xml:space="preserve">
-
-	<t t-name="UOMButton" owl="1">
-		<span class="control-button">
-			<i class='fa fa-tag'></i>
-			<span> </span>
-			<span>UOM</span>
-		</span>
-	</t>
-	
-</templates>
