--- aumet_custom/pos_multi_uom_price/views/pos_multi_uom_price_templates.xml
@@ -1,10 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<odoo>
-	<template id="assets" inherit_id="point_of_sale.assets">
-          <xpath expr="." position="inside">
-	      <script type="text/javascript" src="/pos_multi_uom_price/static/src/js/multi_uom_price.js"></script> 
-	      <script type="text/javascript" src="/pos_multi_uom_price/static/src/js/models.js"></script> 
-          </xpath>
-        </template>
-
-</odoo>
